requirements.txt
requests==2.31.0
PyGithub==1.55